chore: satisfy rustfmt and clippy across qcode and gui
Format the workspace and clear all `clippy -D warnings` findings outside the
Python bindings: collapsible ifs/matches, needless refs, `is_multiple_of`,
`iter().any()`, `let...else`→`?`, doc-list indentation, unused imports, and
same-type casts. Factor complex tuple types into aliases, gate the test-only
`Congruence::congruent` behind `cfg(test)`, allow `too_many_arguments` on the
few wide internal helpers, and `allow(deprecated)` the legacy-ABI
`set_input_regs` calls on the conventional/external path.

+++ analysis/src/gvn/mem_forward.rs
@@ -565,8 +565,8 @@ impl MemForward {
                     CallClobbers::Regs(regs) => !regs.iter().any(|&r| {
                         let vn = Varnode::from_id(ctx, r);
                         vn.space().id == space
-                            && (vn.address() as i64) <= off
-                            && off < vn.address() as i64 + vn.size() as i64
+                            && vn.address() <= off
+                            && off < vn.address() + vn.size() as i64
                     }),
                 },
                 Base::Symbolic(_, bv) => match &clobbers {
